Create Update Delete Growth

+++ Models/Calve.cs
@@ -14,8 +14,8 @@ namespace DairyAPI.Models
         public string cvCalveNo { get; set; }
         public string cvCalveName { get; set; }
         public string cvPostCalveStatus { get; set; }
-        public float? cvWeight { get; set; }
-        public float? cvCalveSalePrice { get; set; }
+        public double? cvWeight { get; set; }
+        public double? cvCalveSalePrice { get; set; }
         public DateTime? date_updated { get; set; }
         public string user_updated { get; set; }
     }
