+++ Models/CowFarmsGrowth.cs
@@ -28,14 +28,14 @@ namespace DairyAPI.Models
         public int? gTranId { get; set; }
         public DateTime? gMeasureDate { get; set; }
         public string gMeasureType { get; set; }
-        public float? gHeartGirth { get; set; }
-        public float? gWeight { get; set; }
+        public double? gHeartGirth { get; set; }
+        public double? gWeight { get; set; }
         public int? gBodyConditionScore { get; set; }
         public string gCowStatus { get; set; }
         public string gEvaluator { get; set; }
         public string gRemark { get; set; }
-        public float? gBodylength { get; set; }
-        public float? gHeight { get; set; }
+        public double? gBodylength { get; set; }
+        public double? gHeight { get; set; }
         public string gTranType { get; set; }
         public DateTime? date_updated { get; set; }
         public string user_updated { get; set; }
